+++ platforms/android/platform_www/plugins/cordova-sqlite-evcore-extbuild-free/www/SQLitePlugin.js
@@ -1,7 +1,7 @@
 cordova.define("cordova-sqlite-evcore-extbuild-free.SQLitePlugin", function(require, exports, module) {
 /*
 License for this version: GPL v3 (http://www.gnu.org/licenses/gpl.txt) or commercial license.
-Contact for commercial license: info@litehelpers.net
+Contact for commercial license: sales@litehelpers.net
  */
 
 (function() {